Minor tweaks
Include space for frame score to be scored in the database with the game
data when inserting scores. This should eventually calculate the score
automatically and placce it in the frame box.

User registration will require email address as well as username to
allow users to recover lost passwords at a later date. This will still
need to be written.