Created a few things in the todo file
Decided this would work well if I just created my about me section as
the home page for the site, and then linked it to the rest of the
bowling admin panel.

If I create a blog I will want to keep the database which includes my
posts separate from the db which controls the bowling statistics so I
don't have to  worry about the security issues.